# src/common/utils/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ficenta API",
    "description": "Tài liệu API cho hệ thống quản lý tài chính cá nhân Ficenta. Tất cả các phản hồi thành công đều được đóng gói trong một cấu trúc chuẩn thông qua Interceptor (success, message, data).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/budget": {
      "get": {
        "tags": [
          "Budget"
        ],
        "summary": "Danh sách ngân sách",
        "description": "Lấy danh sách ngân sách của người dùng theo tháng và năm. Nếu không truyền query param, mặc định lấy tháng năm hiện tại.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "description": "Tháng cần xem (định dạng MM, ví dụ: 07)",
            "schema": {
              "type": "string",
              "example": "07"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "description": "Năm cần xem (định dạng YYYY, ví dụ: 2026)",
            "schema": {
              "type": "string",
              "example": "2026"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/BudgetPopulated"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Budget"
        ],
        "summary": "Tạo ngân sách mới",
        "description": "Tạo một ngân sách mới cho một danh mục trong tháng/năm xác định. Một danh mục chỉ được có tối đa 1 ngân sách trong 1 tháng.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBudgetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Budget"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Ngân sách cho danh mục này trong tháng đã tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/budget/{id}": {
      "put": {
        "tags": [
          "Budget"
        ],
        "summary": "Cập nhật ngân sách",
        "description": "Cập nhật thông tin ngân sách theo ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID của ngân sách cần cập nhật",
            "schema": {
              "type": "string",
              "example": "64a0f823e2003c2a1c0d4575"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Budget"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Bạn không có quyền sửa đổi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ngân sách không tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Budget"
        ],
        "summary": "Xóa ngân sách",
        "description": "Xóa ngân sách theo ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID của ngân sách cần xóa",
            "schema": {
              "type": "string",
              "example": "64a0f823e2003c2a1c0d4575"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Budget"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Bạn không có quyền sửa đổi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ngân sách không tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/transaction": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Danh sách giao dịch",
        "description": "Lấy tất cả các giao dịch của người dùng hiện tại, có thông tin ví và danh mục được populate.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TransactionPopulated"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transaction"
        ],
        "summary": "Tạo giao dịch mới",
        "description": "Tạo giao dịch thu nhập hoặc chi tiêu mới, đồng thời tự động cập nhật số dư của ví.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Transaction"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Lỗi nghiệp vụ (ví hoặc danh mục không hợp lệ, không có quyền, loại không trùng khớp)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ví hoặc danh mục không tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/transaction/{id}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Thông tin giao dịch theo ID",
        "description": "Lấy chi tiết giao dịch theo ID giao dịch.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID giao dịch cần lấy",
            "schema": {
              "type": "string",
              "example": "64a0f823e2003c2a1c0d4576"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TransactionPopulated"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy giao dịch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Transaction"
        ],
        "summary": "Cập nhật giao dịch",
        "description": "Cập nhật chi tiết giao dịch và tự động tính toán lại số dư của ví cũ và ví mới.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID giao dịch cần cập nhật",
            "schema": {
              "type": "string",
              "example": "64a0f823e2003c2a1c0d4576"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Transaction"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Lỗi nghiệp vụ (ví hoặc danh mục không hợp lệ, loại không trùng khớp)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy giao dịch, ví hoặc danh mục",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transaction"
        ],
        "summary": "Xóa giao dịch",
        "description": "Xóa giao dịch theo ID và hoàn lại số dư ví.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID giao dịch cần xóa",
            "schema": {
              "type": "string",
              "example": "64a0f823e2003c2a1c0d4576"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Transaction"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Giao dịch hoặc ví không tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Execute successfully"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Lỗi dữ liệu đầu vào hoặc không tìm thấy bản ghi"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4569"
          },
          "user_id": {
            "type": "string",
            "nullable": true,
            "example": "64a0f823e2003c2a1c0d4567"
          },
          "name": {
            "type": "string",
            "example": "Ăn uống"
          },
          "icon": {
            "type": "string",
            "example": "food-icon"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "expense"
          },
          "is_default": {
            "type": "boolean",
            "example": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-03T18:49:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-03T18:49:00.000Z"
          },
          "__v": {
            "type": "integer",
            "example": 0
          }
        }
      },
...
      "Wallet": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4568"
          },
          "name": {
            "type": "string",
            "example": "Ví Tiền Mặt"
          },
          "user_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4567"
          },
          "type": {
            "type": "string",
            "enum": [
              "cash",
              "bank",
              "credit_card",
              "savings"
            ],
            "example": "cash"
          },
          "currency": {
            "type": "string",
            "example": "VND"
          },
          "initial_balance": {
            "type": "number",
            "example": 500000
          },
          "current_balance": {
            "type": "number",
            "example": 500000
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-03T18:49:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-03T18:49:00.000Z"
          },
          "__v": {
            "type": "integer",
            "example": 0
          }
        }
      },
...
      "Budget": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4575"
          },
          "user_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4567"
          },
          "category_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4569"
          },
          "month": {
            "type": "string",
            "example": "07"
          },
          "year": {
            "type": "string",
            "example": "2026"
          },
          "amount_limit": {
            "type": "number",
            "example": 5000000
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-03T18:49:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-03T18:49:00.000Z"
          },
          "__v": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "BudgetPopulated": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4575"
          },
          "user_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4567"
          },
          "category_id": {
            "$ref": "#/components/schemas/Category"
          },
          "month": {
            "type": "string",
            "example": "07"
          },
          "year": {
            "type": "string",
            "example": "2026"
          },
          "amount_limit": {
            "type": "number",
            "example": 5000000
          },
          "spent": {
            "type": "number",
            "example": 1250000
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-03T18:49:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-03T18:49:00.000Z"
          },
          "__v": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "CreateBudgetDto": {
        "type": "object",
        "required": [
          "category_id",
          "amount_limit",
          "month",
          "year"
        ],
        "properties": {
          "category_id": {
            "type": "string",
            "format": "mongo-id",
            "example": "64a0f823e2003c2a1c0d4569"
          },
          "amount_limit": {
            "type": "number",
            "minimum": 0,
            "example": 5000000
          },
          "month": {
            "type": "string",
            "example": "07"
          },
          "year": {
            "type": "string",
            "example": "2026"
          }
        }
      },
      "UpdateBudgetDto": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "string",
            "format": "mongo-id",
            "example": "64a0f823e2003c2a1c0d4569"
          },
          "amount_limit": {
            "type": "number",
            "minimum": 0,
            "example": 6000000
          },
          "month": {
            "type": "string",
            "example": "07"
          },
          "year": {
            "type": "string",
            "example": "2026"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4576"
          },
          "wallet_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4568"
          },
          "category_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4569"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "expense"
          },
          "amount": {
            "type": "number",
            "example": 150000
          },
          "note": {
            "type": "string",
            "example": "Ăn trưa"
          },
          "transaction_date": {
            "type": "string",
            "example": "2026-07-06"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-06T11:20:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-06T11:20:00.000Z"
          },
          "__v": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "TransactionPopulated": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a0f823e2003c2a1c0d4576"
          },
          "wallet_id": {
            "$ref": "#/components/schemas/Wallet"
          },
          "category_id": {
            "$ref": "#/components/schemas/Category"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "expense"
          },
          "amount": {
            "type": "number",
            "example": 150000
          },
          "note": {
            "type": "string",
            "example": "Ăn trưa"
          },
          "transaction_date": {
            "type": "string",
            "example": "2026-07-06"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-06T11:20:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-06T11:20:00.000Z"
          },
          "__v": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "CreateTransactionDto": {
        "type": "object",
        "required": [
          "wallet_id",
          "category_id",
          "type",
          "amount",
          "transaction_date"
        ],
        "properties": {
          "wallet_id": {
            "type": "string",
            "format": "mongo-id",
            "example": "64a0f823e2003c2a1c0d4568"
          },
          "category_id": {
            "type": "string",
            "format": "mongo-id",
            "example": "64a0f823e2003c2a1c0d4569"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "expense"
          },
          "amount": {
            "type": "number",
            "minimum": 0,
            "example": 150000
          },
          "note": {
            "type": "string",
            "example": "Ăn trưa"
          },
          "transaction_date": {
            "type": "string",
            "example": "2026-07-06"
          }
        }
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "wallet_id": {
            "type": "string",
            "format": "mongo-id",
            "example": "64a0f823e2003c2a1c0d4568"
          },
          "category_id": {
            "type": "string",
            "format": "mongo-id",
            "example": "64a0f823e2003c2a1c0d4569"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "expense"
          },
          "amount": {
            "type": "number",
            "minimum": 0,
            "example": 150000
          },
          "note": {
            "type": "string",
            "example": "Ăn trưa"
          },
          "transaction_date": {
            "type": "string",
            "example": "2026-07-06"
          }
        }
      }
    }
  }
}